fix(mart): 第2回レビュー指摘の反映（横展開漏れ2件＋微修正）
- ranking.vue・統合テストに残っていた「棚割は mart に保持しない」系コメントを
  「集計軸としては保持しない（棚割1のフィルタは対応済み）」へ更新（crosstab.vue と統一）
- OFFSET の int 乗算が残っていた sales 系2箇所（SalesAnalyticsRepository /
  ProductMasterRepository）を long 乗算化（mart 側と同一の修正を Grep で全件展開）
- 0件時の再構築案内を「入力欄の現在値」ではなく「適用済みの検索条件」基準に変更
  （未適用入力で案内が出入りしない）
- Ratio 集約跡の余分な空行を整理

申し送り（任意・非ブロッキングとして両ロール合意）:
- '00000000' 除外の自動回帰テストは次回シード拡充時に追加
- SharePercent のインライン式統一・metricRows の tbody 分割（厳密な rowgroup スコープ）は任意

検証: dotnet build 0警告0エラー / dotnet test 139件全PASS / nuxt typecheck・build PASS。

+++ backend/src/UndeuxSales.Infrastructure/Queries/ProductMasterRepository.cs
@@ -75,7 +75,8 @@ public sealed class ProductMasterRepository
         var parameters = new DynamicParameters();
         ProductMasterFilterSql.AddParameters(filter, parameters);
         parameters.Add("limit", pageSize);
-        parameters.Add("offset", (page - 1) * pageSize);
+        // page は外部入力のため long で乗算し、巨大値でも負の OFFSET（int オーバーフロー）にしない。
+        parameters.Add("offset", (long)(page - 1) * pageSize);
 
         // SKU 統計・代表画像・売上実績（sales_weekly）を 1 商品 1 行に集約する。
         // 売上実績は自然キー（業態×商品記号×品番）で結合する。売上数量は全期間合計、
